+++ example2/lib/calc_wm.dart
@@ -1,17 +1,10 @@
 import 'package:flutter/cupertino.dart';
 import 'package:sbeu_reactive_pattern/reactive/reactive_wm.dart';
 import 'package:sbeu_reactive_pattern/sbeu_reactive_pattern.dart';
-import 'calc_widget.dart';
 
 class CalcWidgetModel extends ReactiveWidgetModel<int> {
-  CalcWidgetModel() {
-    init();
-  }
+  CalcWidgetModel();
 
   final TextEditingController x1 = TextEditingController();
   final TextEditingController x2 = TextEditingController();
-
-  void init() {
-    setWidget(CalcWidget(wm: this));
-  }
 }